Modificaciones de Almacen.java, Bolleria.java, Pan.java y Producto.java Añado los metodos CalcularPrecioTotalEstanteria, eliminarProducto, mostrarInformacion a Almacen.java Añado getters y setters a Producto.java y el metodo mostrarInfo Añado el metodo comprobarDescuento a Bolleria.java

+++ src/g2o/Pan.java
@@ -7,7 +7,7 @@ public class Pan extends Producto
     
     /**********************************************************************************/
 
-    public Pan(int precio, String nombre, String harina) {
+    public Pan(double precio, String nombre, String harina) {
         super(precio, nombre);
         this.harina = harina;
     }
